refactor categories to custom hook
